# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Back End API",
        "description": "Ecommerce-app endpoint",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development API Server"
        }
    ],
    "paths": {
...
        "/get-top-selling": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get top selling products for the authenticated user",
                "description": "Returns a list of top selling products for the authenticated user.",
                "operationId": "getTopSellingProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Failed to fetch top selling products."
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-new-in": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get new in products",
                "description": "Returns a list of new in products.",
                "operationId": "getNewInProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Failed to fetch new in products."
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/toggle-favorite": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Toggle a product's favorite status for the authenticated user",
                "description": "Toggles the favorite status of a product for the authenticated user. Returns success message and status.",
                "operationId": "toggleFavoriteProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID of the product to toggle favorite status",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation - Product removed from favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produk berhasil dihapus dari favorit."
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Successful operation - Product added to favorites",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produk berhasil ditambahkan ke favorit."
                                        },
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error: The product_id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Failed to toggle favorite status."
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-favorite-products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get favorite products for the authenticated user",
                "description": "Returns a list of favorite products for the authenticated user.",
                "operationId": "getFavoriteProducts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get Favorites Products Successfuly"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Failed to fetch favorite products."
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/get-products-byid-category": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products by category ID",
                "description": "Returns a list of products filtered by category ID.",
                "operationId": "getProductsByIdCategory",
                "parameters": [
                    {
                        "name": "query_params",
                        "in": "query",
                        "description": "Category ID to filter products",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get Products Successfuly"
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error: Failed to fetch products by category ID."
                                        },
                                        "error_code": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategoryResource": {
                "title": "CategoryResource Schema",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "products": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "ProductColorsResource": {
                "title": "ProductColorsResource Schema",
                "properties": {
                    "product_id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "rgb": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductImagesResource": {
                "title": "ProductImagesResource Schema",
                "properties": {
                    "url": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "ProductResource Schema",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "category_id": {
                        "type": "integer"
                    },
                    "discounted_price": {
                        "type": "number",
                        "format": "float"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number",
                        "format": "float"
                    },
                    "sales_number": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "$ref": "#/components/schemas/CategoryResource"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImagesResource"
                        }
                    },
                    "colors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductColorsResource"
                        }
                    },
                    "sizes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductSizeResource"
                        }
                    }
                },
                "type": "object"
            },
            "ProductSizeResource": {
                "title": "ProductSizeResource Schema",
                "properties": {
                    "product_id": {
                        "type": "integer"
                    },
                    "size": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}